ENH: Change size, index and offset definitions: use size_t and ptrdiff_t
Instead of unsigned long and long. Adapt surrounding code to use
specific types where possible (i.e., PointIdentifier, CellIdentifier)
but to use size_t and ptrdiff_t where intent is less than crystal
clear.

This commit focuses on changes in the Code/Common directory, and
associated changes in other directories required to build and test
the changes. The remaining directories still need a sweep through
to eliminate inappropriate usage of "long" and "unsigned long".

This commit enables the use of image sizes, offsets and indexes
larger than 2GB on Win64.

For 32-bit Windows, and for 32-bit and 64-bit Linux, this change
should largely be a no-op. The types involved correlate to the
new types being used.

I chose using size_t and ptrdiff_t so that these builds would remain
working as is, including client code, maximizing backwards compatibility.

Code modifications in Testing and Examples are only needed for Win64 builds.
The Testing and Examples would build without code modifications on non-Win64
builds.

COMP:  Fixed compiler error with missing typedefs

Many of the compiler error occured due to missing
typedefs.

+++ Code/Common/itkDefaultDynamicMeshTraits.h
@@ -81,16 +81,16 @@ public:
 
   /** The type to be used to identify a point.  This should be the index type
    * to the PointsContainer. */
-  typedef unsigned long PointIdentifier;
+  typedef size_t PointIdentifier;
 
   /** The type to be used to identify a cell.  This should be the index type
    * to the CellsContainer. */
-  typedef unsigned long CellIdentifier;
+  typedef size_t CellIdentifier;
 
   /** A type that can be used to identifiy individual boundary features on
    * the cells.  Since this will probably be an index into a static array,
    * this will probably never change from an integer setting. */
-  typedef unsigned long CellFeatureIdentifier;
+  typedef size_t CellFeatureIdentifier;
 
   /** The type of point used by the mesh. */
   typedef Point< CoordRepType, VPointDimension > PointType;
